+++ com/baidu/tieba/ala/alasquare/live_tab/view/k.java
@@ -267,10 +267,10 @@ public class k {
 
     public void onChangeSkinType(TbPageContext<?> tbPageContext, int i) {
         if (this.mSkinType != i) {
-            ap.setViewTextColor(this.gtK, R.color.CAM_X0101);
-            ap.setViewTextColor(this.bAG, R.color.CAM_X0101);
-            ap.setViewTextColor(this.gqG, R.color.CAM_X0105);
-            Drawable drawable = ap.getDrawable(this.mTbPageContext.getResources(), R.drawable.tab_icon_living_seeding);
+            ap.setViewTextColor(this.gtK, (int) R.color.CAM_X0101);
+            ap.setViewTextColor(this.bAG, (int) R.color.CAM_X0101);
+            ap.setViewTextColor(this.gqG, (int) R.color.CAM_X0105);
+            Drawable drawable = ap.getDrawable(this.mTbPageContext.getResources(), (int) R.drawable.tab_icon_living_seeding);
             drawable.setBounds(0, 0, this.mTbPageContext.getResources().getDimensionPixelOffset(R.dimen.tbds18), this.mTbPageContext.getResources().getDimensionPixelOffset(R.dimen.tbds18));
             this.gtK.setCompoundDrawablesWithIntrinsicBounds(drawable, (Drawable) null, (Drawable) null, (Drawable) null);
             this.gtK.setCompoundDrawablePadding(this.mTbPageContext.getResources().getDimensionPixelOffset(R.dimen.tbds8));
