+++ com/baidu/tieba/ala/frsgamelive/c/h.java
@@ -116,11 +116,11 @@ public class h implements View.OnClickListener {
     public void onChangeSkinType(TbPageContext<?> tbPageContext, int i) {
         if (this.mSkinType != i) {
             ap.setBackgroundResource(this.gMQ.gzW, R.drawable.addresslist_item_bg);
-            ap.setViewTextColor(this.gMQ.gzU, R.color.CAM_X0111);
-            ap.setViewTextColor(this.gMQ.gzV, R.color.CAM_X0101);
-            ap.setViewTextColor(this.gMQ.aYw, R.color.CAM_X0109);
-            ap.setViewTextColor(this.gMQ.gzY, R.color.CAM_X0105);
-            ap.setViewTextColor(this.gMQ.gMF, R.color.CAM_X0101);
+            ap.setViewTextColor(this.gMQ.gzU, (int) R.color.CAM_X0111);
+            ap.setViewTextColor(this.gMQ.gzV, (int) R.color.CAM_X0101);
+            ap.setViewTextColor(this.gMQ.aYw, (int) R.color.CAM_X0109);
+            ap.setViewTextColor(this.gMQ.gzY, (int) R.color.CAM_X0105);
+            ap.setViewTextColor(this.gMQ.gMF, (int) R.color.CAM_X0101);
             this.mSkinType = i;
         }
     }
